+++ src/scripts/tasks/TaskList.js
@@ -1,12 +1,5 @@
 // Author: Blake Slochowsky
-// Purpose: listen for the delete button event.
-// listnen for the addtaskbuttonclicked event and create a new object, then invoke the save task and pass in the newtask objet in as an argument
-// inside of the render fnction itterate over the array of tasks but before I return the HTML rep. of the task object
-// I make a condition so that it only returns the task whos completed value is set to false. if the value is true then it will ot display it.
-// I also make the event that listens for the click on the checkbox so that when a task is clicked 
-// it finds the FK ID of the task that was clicked matches it with the PK id of the task so it knows which one was completed, then changes that completed value to true causing the task to be hidden.
-// also made a new object that would represent a task completed, then call the updateTask fetch and pass in the new object as an argument that will render our new list of 
-// tasks while taking the task that was selected and hiding it.
+// Purpose: list all of the un completed tasks
 import { useTasks, deleteTask, getTasks, saveTask, updateTask } from "./TaskProvider.js"
 import { TaskComponent } from "./TaskHTML.js"
 import { TaskForm } from "./TaskForm.js"
@@ -35,9 +28,12 @@ eventHub.addEventListener("addTaskButtonClicked", event => {
     saveTask(newTask)
 })
 
+eventHub.addEventListener('userChosen', (event) => {
+    contentTarget.classList.remove("hidden")
+    addButtonTarget.innerHTML = `
+    <button id="addTask--">Add Task </button>`
+})
 
-addButtonTarget.innerHTML = `
-<button id="addTask--">Add Task </button>`
 
 
 eventHub.addEventListener("taskCreated", click => {
@@ -90,6 +86,7 @@ eventHub.addEventListener("taskStateEventChanged", event => {
 
 export const TaskList = () => {
     eventHub.addEventListener("userChosen", (event) => {
+        contentTarget.classList.remove("hidden")
         render()
     })
 }
