+++ src/styles/main.css
@@ -1,3 +1,4 @@
 @import './login.css';
 @import './events.css';
+@import './tasks.css';
 @import './messages/messages.css';
